Extracted `MameSession` code from `src/runtime/controller.rs`
