+++ src/appwindow.rs
@@ -60,10 +60,10 @@ use crate::models::itemstable::ItemsTableModel;
 use crate::prefs::BuiltinCollection;
 use crate::prefs::Preferences;
 use crate::prefs::SortOrder;
-use crate::runtime::controller::MameCommand;
 use crate::runtime::controller::MameController;
-use crate::runtime::controller::MameEvent;
-use crate::runtime::controller::MameStderr;
+use crate::runtime::MameCommand;
+use crate::runtime::MameEvent;
+use crate::runtime::MameStderr;
 use crate::runtime::MameWindowing;
 use crate::selection::SelectionManager;
 use crate::status::Status;
